backend/requirements.txt
ariadne-extended==0.2.2
dj-rest-auth==2.1.4
Django==3.2
django-allauth==0.44.0
django-cors-headers==3.7.0
django-extensions==3.1.2
django-filter==2.4.0
django-waffle==2.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
python-dateutil==2.8.1